First Draft for 1.3
CHANGE : Le niveau des monstres peut désormais dépasser la capacité d'un entier 32 bits.
CHANGE : Le niveau des pièges peut désormais dépasser la capacité d'un entier 32 bits.
CHANGE : C'est une équation qui régit désormais le niveau d'efficacité des objets en fonction du niveau des monstres.
CHANGE : C'est une équation qui régit désormais le niveau d'enchantement des objets en fonction du niveau des monstres.
CHANGE : Le niveau des pièges peut désormais dépasser le niveau de la zone.
CHANGE : Les règles composées peuvent maintenant faire intervenir un nombre arbitraire de clauses.
CHANGE : Refonte du système de haut-faits.
CHANGE : Refonte des objets et du système de butin.

FEATURE : Ajout d'un système de solde et d'une constante définissant la probabilité qu'un objet soit soldés.
FEATURE : Les objets soldés peuvent être achetés à prix cassé, mais sont aussi revendus à prix réduit.
FEATURE : Si un objet soldé est utilisé dans la forge, la sortie sera aussi un objet soldé (pour la plupart des formules).
FEATURE : Ajout d'un système permettant d'automatiser la distribution des points de compétences.
FEATURE : Ajout d'un système de priorité pour les points de compétences, utilisée lors de leur distribution automatique.
FEATURE : Ajout de règles pour filtrer les objets soldés.
FEATURE : Ajout de règles pour filtrer les orbes en fonction de leur nom.
FEATURE : Ajout de règles pour la recherche automatique d'une forge mystique.
FEATURE : Ajout de règles pour automatiser le crafting.
FEATURE : Ajout des compétences "Promophile", "Affinité à la sous-charge" et "Affinité au haut-faits".

FIX : Il n'est plus possible de voir une règle concernant un autre joueur après avoir chargé une sauvegarde.
FIX : La liste des héros est désormais correctement rafraichie quand on charge une sauvegarde.
FIX : L'écran d'édition d'univers pouvait parfois arrêter d'afficher le texte décrivant les équations.

+++ Game.java
@@ -43,7 +43,7 @@ public class Game {
 		Player joueur = new Player(U, false);
 		joueur.disp = false;
 		joueur.defi = ChallengeList.list.get(cli);
-		int previous_level = joueur.level;
+		double previous_level = joueur.level;
 		int defaite_count = 0;
 		int victory_count = 0;
 		int trap_death_count = 0;
